requirements.txt
diffusers==0.16.1
numpy==1.24.3
transformers==4.29.1
xformers==0.0.19
accelerate==0.19.0
Pillow==9.5.0
# Install torch and torchvision manually:
# pip install torch==2.0.0+cu118 torchvision==0.15.1+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
